modified App,Cards,Checkout.js
