+++ src/Cards.js
@@ -1,3 +1,14 @@
+import React from 'react';
+
+const Cards = () => {
+    return (
+        <div>
+            Hello World
+        </div>
+    );
+};
+
+export default Cards;
 
 
 
@@ -21,5 +32,3 @@
 
 
 
-
-export default Cards
